internal(popup): 플러그인 아이콘 클릭시 나오는 popup 창 구조 개선 및 css 정리
